+++ docs/changelog.rst
@@ -5,6 +5,9 @@ CHANGELOG
 2.104.0+dev (XXXX-XX-XX)
 ------------------------
 
+**Bug fixes**
+
+- Set MenuItem `open_in_new_tab` to False during migration when the target is a flat page
 
 
 2.104.0 (2024-04-03)